% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_length}
\alias{chr_length}
\title{Pepper chromosome lengths}
\format{
A GRanges object
}
\usage{
data(chr_length)
}
\description{
Lengths of pepper chromosomes 1-12 in a GRanges object.
The genome for which lengths were calculated (v1.55) was downloaded from
http://peppergenome.snu.ac.kr/download.php
}
\examples{
data(chr_length)
}
\keyword{datasets}
