% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcn}
\alias{gcn}
\title{Simulation of the output list from BioNERO::exp2gcn() with pepper data}
\format{
A list with the elements returned by BioNERO::exp2gcn().
}
\usage{
data(gcn)
}
\description{
This object is a list as returned by BioNERO::exp2gcn(), but only the
element genes_and_modules is included. For running time issues,
only genes in the cyan module were kept in the element genes_and_modules.
All other list elements have been assigned NULL.
The network was inferred using the code from the vignette.
}
\examples{
data(gcn)
}
\keyword{datasets}
