% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_ranges}
\alias{gene_ranges}
\title{Genomic coordinates of pepper genes}
\format{
A GRanges object
}
\usage{
data(gene_ranges)
}
\description{
GRanges object with genomic coordinates of pepper genes downloaded from
http://peppergenome.snu.ac.kr/download.php.
}
\examples{
data(gene_ranges)
}
\keyword{datasets}
