% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_intervals_operations.R
\name{get_all_candidates}
\alias{get_all_candidates}
\title{Get candidate genes for a given sliding window}
\usage{
get_all_candidates(genes_ranges, marker_ranges, window = 2)
}
\arguments{
\item{genes_ranges}{A GRanges object with genomic coordinates
of all genes in the genome.}

\item{marker_ranges}{A GRanges or GRangesList object with
positions of molecular markers.}

\item{window}{Sliding window (in Mb) upstream and downstream relative
to each SNP. Default: 2.}
}
\value{
A GRanges or GRangesList object with the genomic positions of
candidate genes.
}
\description{
For an user-defined sliding window relative to each SNP, this function will
subset all genes whose genomic positions overlap with the sliding window.
}
\examples{
data(snp_pos)
data(gene_ranges)
genes <- get_all_candidates(gene_ranges, snp_pos, window = 2)
}
\seealso{
\code{\link[IRanges]{findOverlaps-methods}}
}
