% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guides}
\alias{guides}
\title{Guide genes associated with sucrose metabolism}
\format{
A data frame with genes in the first column and MapMan in the
second column.
}
\usage{
guides
}
\description{
The MapMan annotation was downloaded from PLAZA Monocots.
}
\examples{
data(guides)
}
\references{
Van Bel, M., Diels, T., Vancaester, E., Kreft, L., Botzki, A.,
Van de Peer, Y., ... & Vandepoele, K. (2018). PLAZA 4.0: an integrative
resource for functional, evolutionary and comparative plant
genomics. Nucleic acids research, 46(D1), D1190-D1196.
}
\keyword{datasets}
