% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas}
\alias{gwas}
\title{GWAS-derived SNPs for maize traits}
\format{
A GRanges object.
}
\usage{
data(gwas)
}
\description{
The SNPs in this data set were retrieved from Wallace et al., 2014, and
they are associated to chlorophyll A, chlorophyll B, nitrate and sucrose.
The SNP positions and associated traits are stored in a GRanges object.
}
\examples{
data(gwas)
}
\references{
Wallace JG, Bradbury PJ, Zhang N, Gibon Y, Stitt M, Buckler ES (2014)
Association Mapping across Numerous Traits Reveals Patterns of Functional
Variation in Maize. PLoS Genet 10(12): e1004845.
doi:10.1371/journal.pgen.1004845
}
\keyword{datasets}
