% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hubs}
\alias{hubs}
\title{Example hub genes for the network stored in the gcn object}
\format{
Data frame with gene IDs, module and intramodular degree.
}
\usage{
data(hubs)
}
\description{
The data frame was created using the code from the vignette.
}
\examples{
data(hubs)
}
\keyword{datasets}
