% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_exp}
\alias{maize_exp}
\title{Maize expression data from Stelpflug et al., 2016}
\format{
A SummarizedExperiment object.
}
\usage{
maize_exp
}
\description{
The data were filtered to remove genes with median expression levels <2
for package size issues.
}
\examples{
data(maize_exp)
}
\references{
Stelpflug, S.C., Sekhon, R.S., Vaillancourt, B., Hirsch, C.N., Buell,
C.R., de Leon, N. and Kaeppler, S.M. (2016), An Expanded Maize Gene
Expression Atlas based on RNA Sequencing and its Use to Explore Root
Development. The Plant Genome, 9: plantgenome2015.04.0025.
https://doi.org/10.3835/plantgenome2015.04.0025
}
\keyword{datasets}
