% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maize_gr}
\alias{maize_gr}
\title{Genomic coordinates for maize features}
\format{
A GRanges object
}
\usage{
data(maize_gr)
}
\description{
The genomic coordinates for maize features (e.g., genes, transcripts)
are stored in a GRanges object. Some columns of the .gff file were considered
unnecessary and, hence, they were removed to reduce package size. The .gff
file was obtained from PLAZA 4.0 Monocots.
}
\examples{
data(maize_gr)
}
\references{
Van Bel, M., Diels, T., Vancaester, E., Kreft, L., Botzki, A.,
Van de Peer, Y., ... & Vandepoele, K. (2018). PLAZA 4.0: an integrative
resource for functional, evolutionary and comparative plant
genomics. Nucleic acids research, 46(D1), D1190-D1196.
}
\keyword{datasets}
