% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mine2}
\alias{mine2}
\title{Example output from mine_step2()}
\format{
List with elements 'candidates' (character vector) and 'enrichment'
(data frame).
}
\usage{
data(mine2)
}
\description{
The list was created using the example code from mine_step().
}
\examples{
data(mine2)
}
\keyword{datasets}
