% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_mining.R
\name{mine_candidates}
\alias{mine_candidates}
\title{Mine high-confidence candidate genes}
\usage{
mine_candidates(
  exp,
  metadata,
  gcn,
  guides,
  candidates,
  min_cor = 0.2,
  sample_group,
  alpha = 0.05,
  continuous = FALSE
)
}
\arguments{
\item{exp}{Expression data frame with genes in row names and samples in
column names or a SummarizedExperiment object.}

\item{metadata}{Sample metadata with samples in row names and sample
information in the first column. Ignored if \code{exp} is a SummarizedExperiment
object, as the colData will be extracted from the object.}

\item{gcn}{Gene coexpression network returned by \code{BioNERO::exp2gcn()}.}

\item{guides}{Guide genes as a character vector or as a data frame with
genes in the first column and gene annotation class in the second column.}

\item{candidates}{Character vector of all candidates genes to be inspected.}

\item{min_cor}{Minimum correlation value for
\code{BioNERO::gene_significance()}. Default: 0.2}

\item{sample_group}{Level of sample metadata to be used for filtering
in gene-trait correlation.}

\item{alpha}{Numeric indicating significance level. Default: 0.05}

\item{continuous}{Logical indicating whether metadata is continuous or not.
Default: FALSE}
}
\value{
A data frame with high-confidence genes, their correlations to
\code{sample_group} and correlation p-values.
}
\description{
Mine high-confidence candidate genes
}
\details{
DETAILS
}
\examples{
data(maize_exp)
data(gwas)
data(maize_gr)
data(guides)
genes_ranges <- maize_gr[maize_gr$type == "gene", ]
marker_ranges <- gwas[gwas$trait == "Sucrose"]
genes <- get_all_candidates(genes_ranges, marker_ranges, window = 2)$ID
exp <- maize_exp
exp <- BioNERO::exp_preprocess(exp, cor_method = "pearson", min_exp=5)
# SFT power previously calculated
gcn <- BioNERO::exp2gcn(exp, net_type = "signed", cor_method = "pearson",
                        module_merging_threshold = 0.8, SFTpower = 12)
}
\seealso{
\code{BioNERO::module_enrichment},
\code{BioNERO::gene_significance}
}
