% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mined_candidates}
\alias{mined_candidates}
\title{Example output from mined_candidates()}
\format{
Data frame with an example of the output from mined_candidates
}
\usage{
data(mined_candidates)
}
\description{
The data frame was created using the code from the vignette.
}
\examples{
data(mined_candidates)
}
\keyword{datasets}
