% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepper_se}
\alias{pepper_se}
\title{Gene expression data from Kim et al., 2018.}
\format{
A SummarizedExperiment object.
}
\usage{
pepper_se
}
\description{
The data were filtered to keep only the top 4000 genes with highest RPKM
values in PRR stress-related samples.
}
\examples{
data(pepper_se)
}
\references{
Kim, MS., Kim, S., Jeon, J. et al. Global gene expression profiling for
fruit organs and pathogen infections in the pepper, Capsicum annuum L..
Sci Data 5, 180103 (2018). https://doi.org/10.1038/sdata.2018.103
}
\keyword{datasets}
