% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_snp_distribution}
\alias{plot_snp_distribution}
\title{Plot SNP distribution across chromosomes}
\usage{
plot_snp_distribution(marker_ranges)
}
\arguments{
\item{marker_ranges}{A GRanges object with genomic positions
of molecular markers.}
}
\value{
A ggplot object.
}
\description{
Plot SNP distribution across chromosomes
}
\examples{
data(gwas)
p <- plot_snp_distribution(gwas)
}
