% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_pos}
\alias{snp_pos}
\title{Capsicum annuum SNPs associated with resistance to Phytophthora root rot.}
\format{
A GRanges object.
}
\usage{
data(snp_pos)
}
\description{
The SNPs in this data set were retrieved from Siddique et al., 2019, and
they are associated to resistance to Phytophthora root rot.
}
\examples{
data(snp_pos)
}
\references{
Siddique, M.I., Lee, HY., Ro, NY. et al. Identifying candidate genes for
Phytophthora capsici resistance in pepper (Capsicum annuum) via
genotyping-by-sequencing-based QTL mapping and genome-wide association
study. Sci Rep 9, 9962 (2019). https://doi.org/10.1038/s41598-019-46342-1
}
\keyword{datasets}
