% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfs}
\alias{tfs}
\title{Pepper transcription factors}
\format{
A data frame with gene IDs in the first column and TF families in
the second column.
}
\usage{
data(tfs)
}
\description{
Pepper transcription factors and their families retrieved from PlantTFDB 4.0.
}
\examples{
data(tfs)
}
\references{
Jin, J., Tian, F., Yang, D. C., Meng, Y. Q., Kong, L., Luo, J., &
Gao, G. (2016). PlantTFDB 4.0: toward a central hub for transcription
factors and regulatory interactions in plants. Nucleic acids research, gkw982.
}
\keyword{datasets}
