% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_mining.R
\name{mine_candidates}
\alias{mine_candidates}
\title{Mine high-confidence candidate genes}
\usage{
mine_candidates(
  exp,
  metadata,
  gcn,
  guides,
  candidates,
  sample_group,
  min_cor = 0.2,
  alpha = 0.05,
  continuous = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{exp}{Expression data frame with genes in row names and samples in
column names or a SummarizedExperiment object.}

\item{metadata}{Sample metadata with samples in row names and sample
information in the first column. Ignored if \code{exp} is a SummarizedExperiment
object, as the colData will be extracted from the object.}

\item{gcn}{Gene coexpression network returned by \code{BioNERO::exp2gcn()}.}

\item{guides}{Guide genes as a character vector or as a data frame with
genes in the first column and gene annotation class in the second column.}

\item{candidates}{Character vector of all candidates genes to be inspected.}

\item{sample_group}{Level of sample metadata to be used for filtering
in gene-trait correlation.}

\item{min_cor}{Minimum correlation value for
\code{BioNERO::gene_significance()}. Default: 0.2}

\item{alpha}{Numeric indicating significance level. Default: 0.05}

\item{continuous}{Logical indicating whether metadata is continuous or not.
Default: FALSE}

\item{verbose}{Logical indicating the verbosity of the function.
Default: TRUE}
}
\value{
A data frame with high-confidence genes, their correlations to
\code{sample_group} and correlation p-values.
}
\description{
Mine high-confidence candidate genes
}
\details{
DETAILS
}
\examples{
data(pepper_se)
data(snp_pos)
data(gene_ranges)
data(guides)
set.seed(1)
# Previously selected power = 12
sft <- BioNERO::SFT_fit(pepper_se, net_type = "signed",
                        cor_method = "pearson")
gcn <- BioNERO::exp2gcn(pepper_se, net_type = "signed", cor_method = "pearson",
                        module_merging_threshold = 0.8, SFTpower = 12)
hc_genes <- mine_candidates(pepper_se, gcn = gcn, guides = guides$Gene,
                            candidates = rownames(pepper_se),
                            sample_group = "PRR_stress")
}
\seealso{
\code{BioNERO::module_enrichment},
\code{BioNERO::gene_significance}
}
