% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_snp_circos}
\alias{plot_snp_circos}
\title{Circos plot of molecular marker distribution across chromosomes}
\usage{
plot_snp_circos(genome_ranges, genes_ranges, marker_ranges)
}
\arguments{
\item{genome_ranges}{A GRanges object with chromosome lengths.}

\item{genes_ranges}{A GRanges object with genomic coordinates
of all genes in the genome.}

\item{marker_ranges}{A GRanges or GRangesList object with
positions of molecular markers.}
}
\value{
A ggplot object with a circos plot of molecular marker distribution
across chromosomes.
}
\description{
Circos plot of molecular marker distribution across chromosomes
}
\examples{
data(gwas)
data(maize_gr)
data(chr_length)
gwas_list <- split(gwas, gwas$trait)
genes_ranges <- maize_gr[maize_gr$type == "gene", ]
p <- plot_snp_circos(chr_length, genes_ranges, gwas_list)
}
\seealso{
\code{\link[ggplotify]{as.ggplot}}
\code{\link[ggtext]{element_textbox}}
}
