% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_snp_distribution}
\alias{plot_snp_distribution}
\title{Plot SNP distribution across chromosomes}
\usage{
plot_snp_distribution(marker_ranges, trait_col = NULL)
}
\arguments{
\item{marker_ranges}{A GRanges object with genomic positions.}

\item{trait_col}{Character or numeric index of column containing
trait information. If NULL, a single trait will be implicitly considered.
Default: NULL.
of molecular markers.}
}
\value{
A ggplot object.
}
\description{
Plot SNP distribution across chromosomes
}
\examples{
data(snp_pos)
p <- plot_snp_distribution(snp_pos)
}
