% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_snp_distribution}
\alias{plot_snp_distribution}
\title{Plot SNP distribution across chromosomes}
\usage{
plot_snp_distribution(marker_ranges)
}
\arguments{
\item{marker_ranges}{Genomic positions of SNPs. For a single trait,
a GRanges object. For multiple traits, a GRangesList or CompressedGRangesList
object, with each element of the list representing SNP positions for a
particular trait.}
}
\value{
A ggplot object.
}
\description{
Plot SNP distribution across chromosomes
}
\examples{
data(snp_pos)
p <- plot_snp_distribution(snp_pos)
}
