% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_mining.R
\name{score_genes}
\alias{score_genes}
\title{Score candidate genes and select the top}
\usage{
score_genes(mined_candidates, hubs = NULL, tfs = NULL, pick_top = 10)
}
\arguments{
\item{mined_candidates}{Data frame resulting from \code{mine_candidates}.}

\item{hubs}{Character vector of hub genes.}

\item{tfs}{Character vector of transcription factors.}

\item{pick_top}{Number of top genes to select. Default: 10.}
}
\value{
Data frame with top n candidates.
}
\description{
Score candidate genes and select the top
}
\examples{
data(pepper_se)
data(snp_pos)
data(gene_ranges)
data(guides)
data(tfs)
set.seed(1)
# sft <- BioNERO::SFT_fit(pepper_se, net_type = "signed",
#                         cor_method = "pearson")
# Previously selected power = 12
gcn <- BioNERO::exp2gcn(pepper_se, net_type = "signed", cor_method = "pearson",
                        module_merging_threshold = 0.8, SFTpower = 12)
candidates <- mine_candidates(gene_ranges, snp_pos, exp = pepper_se,
                              gcn = gcn, guides = guides,
                              sample_group = "PRR_stress")
hubs <- BioNERO::get_hubs_gcn(pepper_se, gcn)
scored <- score_genes(candidates, hubs$Gene, tfs$Gene_ID)
}
