% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_intervals_operations.R
\name{simulate_windows}
\alias{simulate_windows}
\title{Simulate number of genes for each sliding window}
\usage{
simulate_windows(genes_ranges, marker_ranges, windows = seq(0.1, 2, by = 0.1))
}
\arguments{
\item{genes_ranges}{A GRanges object with genomic coordinates
of all genes in the genome.}

\item{marker_ranges}{A GRanges or GRangesList object with
positions of molecular markers.}

\item{windows}{Sliding windows (in Mb) upstream and downstream relative
to each SNP. Default: seq(0.1, 2, by = 0.1).}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function counts genes that are contained in sliding windows related to
each SNP.
}
\details{
By default, the function creates 20 sliding windows by expanding upstream
and downstream boundaries for each SNP from 0.1 Mb (100 kb) to 2 Mb.
}
\examples{
data(gwas)
data(maize_gr)
genes_ranges <- maize_gr[maize_gr$type == "gene", ]
marker_ranges <- split(gwas, gwas$trait)
simulate_windows(genes_ranges, marker_ranges)
}
\seealso{
\code{\link[IRanges]{findOverlaps-methods}}
}
